examples/email-subject-scorer/requirements.txt
google-genai==1.14.0
